Redesign Laser into a Particle System. And completely fixed fire anim bug on 92fs
The new particle laser is not only better lookign and performs in the dark. It requires less scripting checks to do. Doubtful it would be a performance increase since the particle emissions are going to be checked per frame as well. But still an improvement visually, also has the side effect of perfectly following the laser base!